[#23] Add missing word "in" and quote library path.
The second to last installation step is missing the word "in"
before " your <arduinosketchfolder>/libraries/ folder,".
The "<arduinosketchfolder>" will not show properly in rendered
Markdown because  '<' is the start of an HTML tag. Enclosing this
in backticks fixes it. Alternatively, the less than and greater than
symbols could be HTML encoded, but this does not look as clean in the
raw Markdown text.